pyproject.toml
[tool.poetry]
name = "employee-hierarchy"
version = "0.1.0"
description = "test assignments"
authors = ["NikolayZemelko <zemelko1989@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.7"
django-extensions = "^3.2.3"
flake8 = "^6.1.0"
django-bootstrap5 = "^23.3"
psycopg = "^3.1.12"
django-filter = "^23.3"
gunicorn = "^21.2.0"
python-dotenv = "0.21.1"
django-crispy-forms = "^2.1"
crispy-bootstrap4 = "^2023.1"

[tool.poetry.group.dev.dependencies]
Faker = "^19.13.0"
django-seeder = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
